Significantly speed up shell boot time 🚅
I noticed shell boot time and `fzf` when using `rg --files` as the
`FZF_DEFAULT_COMMAND` was super slow. I noticed that when opening a new
terminal tab, `ruby` would run for a bit, which didn't seem ideal.
Eventually I traced that down to this. Removing the call out to homebrew
is less flexible but worth it for the significantly faster boot time.